Add removers to AssessmentOption
Added a new SessionTypesEntity trail for managing session types
relationships.  This trait is used in AamcMethod and School as well.
